+++ changelog.md
@@ -1,2 +1,8 @@
 v1.0.0:
     - Initial
+v1.0.1:
+    - Added React Router
+    - Added Routes
+    - Added NavBar
+    - Added Login page
+    - Added basic pages in pages folder